Chore: 홈페이지 컴포넌트화
