+++ src/components/Home/HomeNavbar.jsx
@@ -2,6 +2,8 @@ import Image from 'next/image';
 import Link from 'next/link';
 
 export default function HomeNavbar() {
+  const hoverStyle =
+    'hover:border-b-2 hover:border-pink-400 hover:text-pink-400 border-b-2 border-transparent t pb-1 transition-all';
   return (
     <>
       <header className='h-[78px] bg-white'>
@@ -13,8 +15,12 @@ export default function HomeNavbar() {
           </figure>
           <ul className='flex w-full justify-between gap-4 text-[20px] font-bold'>
             <div className='mx-auto flex gap-4'>
-              <li className='border-b-2 border-pink-400 pb-1 text-pink-400'>장소찾기</li>
-              <li>커뮤니티</li>
+              <Link href='/content/place'>
+                <li className={`${hoverStyle} `}>장소찾기</li>
+              </Link>
+              <Link href='/content/community'>
+                <li className={hoverStyle}>커뮤니티</li>
+              </Link>
             </div>
             <Link href='/login' className='mr-[50px]'>
               로그인
